Migrated plugin/property to unsupported/property-old and unsupported/property-ng to plugin/property. Updated poms. Refactored packages. Updated display names
