+++ modules/plugin/property/src/main/java/org/geotools/data/property/PropertyFeatureSource.java
@@ -16,74 +16,145 @@
  */
 package org.geotools.data.property;
 
+import java.io.BufferedReader;
 import java.io.File;
+import java.io.FileReader;
 import java.io.IOException;
 
-import org.geotools.data.AbstractFeatureSource;
-import org.geotools.data.FeatureListener;
+import org.geotools.data.DataSourceException;
+import org.geotools.data.DataUtilities;
+import org.geotools.data.FeatureReader;
+import org.geotools.data.Query;
 import org.geotools.data.QueryCapabilities;
+import org.geotools.data.store.ContentEntry;
+import org.geotools.data.store.ContentFeatureSource;
 import org.geotools.factory.Hints;
+import org.geotools.feature.SchemaException;
 import org.geotools.geometry.jts.ReferencedEnvelope;
+import org.geotools.geometry.jts.WKTReader2;
+import org.opengis.feature.FeatureVisitor;
+import org.opengis.feature.simple.SimpleFeature;
 import org.opengis.feature.simple.SimpleFeatureType;
+import org.opengis.filter.Filter;
 
 /**
+ * @author Jody Garnett
+ * @author Torben Barsballe (Boundless)
  * 
- *
  * @source $URL$
  */
-public class PropertyFeatureSource extends AbstractFeatureSource {
+public class PropertyFeatureSource extends ContentFeatureSource {
     String typeName;
     SimpleFeatureType featureType;
     PropertyDataStore store;
-    long cacheTimestamp = 0;
-    ReferencedEnvelope cacheBounds = null;
-
-    PropertyFeatureSource(PropertyDataStore propertyDataStore, String typeName)
-            throws IOException {
-        this.store = propertyDataStore;
-        this.typeName = typeName;
-        this.featureType = store.getSchema(typeName);
-        this.queryCapabilities = new QueryCapabilities() {
+    
+    public PropertyFeatureSource(ContentEntry entry, Query query) {
+        super(entry, query);
+        this.store = (PropertyDataStore) entry.getDataStore();
+        this.typeName = entry.getTypeName();
+    }
+    
+    
+    public PropertyDataStore getDataStore() {
+        return (PropertyDataStore) super.getDataStore();
+    }
+    
+    @Override
+    protected QueryCapabilities buildQueryCapabilities() {
+        return new QueryCapabilities(){
             public boolean isUseProvidedFIDSupported() {
                 return true;
             }
         };
-        hints.add(Hints.FEATURE_DETACHED);
     }
-
-    public PropertyDataStore getDataStore() {
-        return store;
-    }
-
-    public void addFeatureListener(FeatureListener listener) {
-        store.listenerManager.addFeatureListener(this, listener);
-    }
-
-    public void removeFeatureListener(FeatureListener listener) {
-        store.listenerManager.removeFeatureListener(this, listener);
-    }
-
-    public SimpleFeatureType getSchema() {
-        return featureType;
-    }
-
-    // constructor end
-    // getBounds start
-    public ReferencedEnvelope getBounds() {
-        File file = new File(store.directory, typeName + ".properties");
-        if (cacheBounds != null && file.lastModified() == cacheTimestamp) {
-            // we have the cache
-            return cacheBounds;
+    
+    @Override
+    protected ReferencedEnvelope getBoundsInternal(Query query) throws IOException {
+        if (query.getFilter() == Filter.INCLUDE) { //filtering not implemented
+            ReferencedEnvelope bounds = ReferencedEnvelope.create( 
+                    getSchema().getCoordinateReferenceSystem() ); 
+            FeatureReader<SimpleFeatureType, SimpleFeature> featureReader = getReaderInternal(query);
+            try {
+                while (featureReader.hasNext()) {
+                    SimpleFeature feature = featureReader.next();
+                    bounds.include(feature.getBounds());
+                }
+            } finally {
+                featureReader.close();
+            }
+            return bounds;
         }
+        return null; // feature by feature scan required to count records
+    }
+    
+    @Override
+    protected int getCountInternal(Query query) throws IOException {
+        if (query.getFilter() == Filter.INCLUDE) { //filtering not implemented
+            int count = 0;
+            FeatureReader<SimpleFeatureType, SimpleFeature> featureReader = getReaderInternal(query);
+            try {
+                while (featureReader.hasNext()) {
+                    featureReader.next();
+                    count++;
+                }
+            } finally {
+                featureReader.close();
+            }
+            return count;
+        }
+        return -1; // feature by feature scan required to count records
+    }
+    
+    @Override
+    protected SimpleFeatureType buildFeatureType() throws IOException {
+        String typeName = getEntry().getTypeName();
+        String namespace = getEntry().getName().getNamespaceURI();
+        
+        String typeSpec = property("_");
         try {
-            // calculate and store in cache
-            cacheBounds = getFeatures().getBounds();
-            cacheTimestamp = file.lastModified();
-            return cacheBounds;
-        } catch (IOException e) {
+            return DataUtilities.createType(namespace, typeName, typeSpec);
+        } catch (SchemaException e) {
+            e.printStackTrace();
+            throw new DataSourceException(typeName + " schema not available", e);
         }
-        // bounds are unavailable!
+    }
+    
+    private String property(String key) throws IOException {
+        File file = new File( store.dir, typeName+".properties");
+        
+        BufferedReader reader = new BufferedReader(new FileReader(file));
+        try {
+            for (String line = reader.readLine(); line != null; line = reader.readLine()) {
+                if (line.startsWith(key + "=")) {
+                    return line.substring(key.length() + 1);
+                }
+            }
+        } finally {
+            reader.close();
+        }
         return null;
     }
-    // getBounds end
+    
+    @Override
+    protected FeatureReader<SimpleFeatureType, SimpleFeature> getReaderInternal(Query query)
+            throws IOException {
+        File file = new File( store.dir, typeName+".properties");
+        PropertyFeatureReader reader = new PropertyFeatureReader(store.getNamespaceURI(), file);
+        
+        Double tolerance = (Double)query.getHints().get(Hints.LINEARIZATION_TOLERANCE);
+        if (tolerance != null) {
+            reader.setWKTReader(new WKTReader2(tolerance));
+        }
+        
+        return reader;
+    }
+    
+    /**
+     * Make handleVisitor package visible allowing PropertyFeatureStore to delegate to
+     * this implementation.
+     */
+    @Override
+    protected boolean handleVisitor(Query query, FeatureVisitor visitor) throws IOException {
+        return super.handleVisitor(query, visitor);
+    }
 }
