+++ modules/plugin/property/src/test/java/org/geotools/data/property/PropertyDataStore2Test.java
@@ -2,7 +2,7 @@
  *    GeoTools - The Open Source Java GIS Toolkit
  *    http://geotools.org
  *
- *    (C) 2002-2008, Open Source Geospatial Foundation (OSGeo)
+ *    (C) 2002-2014, Open Source Geospatial Foundation (OSGeo)
  *
  *    This library is free software; you can redistribute it and/or
  *    modify it under the terms of the GNU Lesser General Public
@@ -17,78 +17,65 @@
 package org.geotools.data.property;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNotSame;
 import static org.junit.Assert.assertTrue;
 
-import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileNotFoundException;
-import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
-import org.geotools.data.DataUtilities;
+import junit.framework.TestCase;
+
+import org.geotools.data.DataStore;
 import org.geotools.data.Query;
-import org.geotools.data.collection.ListFeatureCollection;
+import org.geotools.data.property.PropertyDataStore;
 import org.geotools.data.simple.SimpleFeatureCollection;
 import org.geotools.data.simple.SimpleFeatureIterator;
 import org.geotools.data.simple.SimpleFeatureSource;
-import org.geotools.data.simple.SimpleFeatureStore;
-import org.geotools.factory.CommonFactoryFinder;
-import org.geotools.factory.Hints;
-import org.geotools.feature.simple.SimpleFeatureBuilder;
-import org.geotools.geometry.jts.JTSFactoryFinder;
+import org.geotools.filter.text.cql2.CQL;
+import org.geotools.geometry.jts.ReferencedEnvelope;
+import org.geotools.geometry.jts.ReferencedEnvelope3D;
 import org.geotools.referencing.CRS;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
-import org.opengis.feature.Feature;
-import org.opengis.feature.FeatureVisitor;
 import org.opengis.feature.Property;
 import org.opengis.feature.simple.SimpleFeature;
 import org.opengis.feature.simple.SimpleFeatureType;
-import org.opengis.feature.type.GeometryDescriptor;
 import org.opengis.feature.type.GeometryType;
 import org.opengis.filter.Filter;
-import org.opengis.filter.FilterFactory2;
 import org.opengis.referencing.crs.CoordinateReferenceSystem;
 
-import com.vividsolutions.jts.geom.Coordinate;
 import com.vividsolutions.jts.geom.Geometry;
-import com.vividsolutions.jts.geom.GeometryFactory;
-import com.vividsolutions.jts.geom.LineString;
 
 /**
- * Test non functionality of PropertyDataStore (problems and enhancements requested through JIRA).
+ * Test non functionality of PropertyDataStore.
  * 
- * @author Jody Garnett (LISAsoft)
- *
+ * @author Jody Garnett, Refractions Research Inc.
+ * @author Torben Barsballe (Boundless)
  *
  * @source $URL$
  */
-public class PropertyDataStore2Test {
+public class PropertyDataStore2Test extends TestCase {
     PropertyDataStore store;
-    
+    PropertyDataStore store3d;
     PropertyDataStore sridStore;
-
-    private PropertyDataStore unusalStore;
-
-    @Before
-    public void setUp() throws Exception {
+    
+    /**
+     * Constructor for SimpleDataStoreTest.
+     * @param arg0
+     */
+    public PropertyDataStore2Test(String arg0) {
+        super(arg0);
+    }
+    protected void setUp() throws Exception {
         File dir = new File(".", "propertyTestData" );
         dir.mkdir();
-               
+        
         File file = new File( dir ,"road.properties");
-        if( file.exists()){
+        if (file.exists()) {
             file.delete();
-        }        
+        }
         BufferedWriter writer = new BufferedWriter( new FileWriter( file ) );
         writer.write("_=id:Integer,*geom:Geometry,name:String"); writer.newLine();
         writer.write("fid1=1|LINESTRING(0 0,10 10)|jody"); writer.newLine();
@@ -96,7 +83,7 @@ public class PropertyDataStore2Test {
         writer.write("fid3=3|LINESTRING(5 0, 5 10)|dave"); writer.newLine();
         writer.write("fid4=4|LINESTRING(0 5, 5 0, 10 5, 5 10, 0 5)|justin");
         writer.close();
-        store = new PropertyDataStore( dir );
+        store = new PropertyDataStore( dir, "propertyTestData" );
         
         // Create a similar data store but with srid in the geometry column
         File dir2 = new File(".", "propertyTestData2");
@@ -116,48 +103,50 @@ public class PropertyDataStore2Test {
         writer2.newLine();
         writer2.write("fid4=4|LINESTRING(0 5, 5 0, 10 5, 5 10, 0 5)");
         writer2.close();
-        sridStore = new PropertyDataStore(dir2);
-
-        // Create a similar data store with various unusal problems
+        sridStore = new PropertyDataStore( dir2, "propertyTestData2" );
+        
+        // Create a data store with a 3D geometry
         File dir3 = new File(".", "propertyTestData3");
         dir3.mkdir();
-        File file3 = new File( dir3 ,"unusual.properties");
-        if( file3.exists()){
+        File file3 = new File(dir3, "road3.properties");
+        if (file3.exists()) {
             file3.delete();
         }
         
-        BufferedWriter writer3 = new BufferedWriter( new FileWriter( file3 ) );
-        writer3.write("_=id:Integer,*geom:Geometry,data:String"); writer3.newLine();
-        writer3.write("fid1=1|LINESTRING(0 0,10 10)|feeling a bit \\|broken up"); writer3.newLine();
-        writer3.write("fid2=2|\\\nLINESTRING(20 20,30 30)|\\\nnew line\\\ndiff friendly");writer3.newLine();
+        BufferedWriter writer3 = new BufferedWriter(new FileWriter(file3));
+        writer3.write("_=id:Integer,geom:Geometry:srid=4327");
+        writer3.newLine();
+        writer3.write("fid1=1|LINESTRING(0 0 0,10 10 0)");
+        writer3.newLine();
+        writer3.write("fid2=2|LINESTRING(20 20 10,30 30 20)");
         writer3.close();
-        unusalStore = new PropertyDataStore( dir3 );
+        store3d = new PropertyDataStore( dir3, "propertyTestData3" );
         
-        // listFileContents( file3 );
+        super.setUp();
     }
-    
-    @After
-    public void tearDown() throws Exception {
-        store.dispose();
+    protected void tearDown() throws Exception {
         File dir = new File( "propertyTestData" );
-        for( File file : dir.listFiles()){
-            file.delete();
+        File list[]=dir.listFiles();
+        for( int i=0; i<list.length;i++){
+            list[i].delete();
         }
         dir.delete();
-
-        sridStore.dispose();
+        
         dir = new File( "propertyTestData2" );
-        for( File file : dir.listFiles()){
-            file.delete();
+        list = dir.listFiles();
+        for( int i=0; i<list.length;i++){
+            list[i].delete();
         }
         dir.delete();
-
-        unusalStore.dispose();
+        
         dir = new File( "propertyTestData3" );
-        for( File file : dir.listFiles()){
-            file.delete();
+        list = dir.listFiles();
+        for( int i=0; i<list.length;i++){
+            list[i].delete();
         }
         dir.delete();
+        
+        super.tearDown();
     }
     
     /**
@@ -165,12 +154,11 @@ public class PropertyDataStore2Test {
      * 
      * @throws Exception
      */
-    @Test
     public void testCRS() throws Exception {
         SimpleFeatureSource road = sridStore.getFeatureSource("road2");
         SimpleFeatureCollection features = road.getFeatures();
         assertEquals(4, features.size());
-
+        
         SimpleFeature feature;
         Geometry geom;
         Property prop;
@@ -181,48 +169,39 @@ public class PropertyDataStore2Test {
             prop = feature.getProperty("geom");
             assertTrue(prop.getType() instanceof GeometryType);
             geomType = (GeometryType) prop.getType();
-
+            
             Object val = prop.getValue();
             assertTrue(val != null && val instanceof Geometry);
             geom = (Geometry) val;
-
+            
             Object userData = geom.getUserData();
             assertTrue(userData != null && userData instanceof CoordinateReferenceSystem);
             // ensure the same CRS is passed on to userData for encoding
             assertEquals(userData, geomType.getCoordinateReferenceSystem());
         }
     }
-    @Test
-    public void unusual() throws Exception {
-        SimpleFeatureSource road = unusalStore.getFeatureSource( "unusual" );
-        assertEquals( 3, road.getSchema().getAttributeCount() );
-
-        SimpleFeatureCollection features = road.getFeatures();
-        assertFalse( features.isEmpty() );
+    
+    public void test3D() throws Exception {
+        SimpleFeatureSource fs = store3d.getFeatureSource("road3");
+        Query q = new Query( "road3", Filter.INCLUDE);
+        assertEquals(2, fs.getCount(q));
+        ReferencedEnvelope bounds = new ReferencedEnvelope3D(0, 30, 0, 30, 0, 20, 
+                fs.getSchema().getCoordinateReferenceSystem());
+        assertEquals(bounds, fs.getBounds());
+        assertEquals(bounds, fs.getBounds(q));
         
-        FilterFactory2 ff = CommonFactoryFinder.getFilterFactory2();
+        SimpleFeatureCollection features = fs.getFeatures();
+        assertEquals(2, features.size());
         
-        SimpleFeatureCollection select = road.getFeatures( ff.id( ff.featureId("fid1")));
-        SimpleFeatureIterator iterator = select.features();
-        SimpleFeature feature1 = iterator.next();
-        assertNotNull( feature1 );
-        iterator.close();
-
-        select = road.getFeatures( ff.id( ff.featureId("fid2")));
-        iterator = select.features();
-        SimpleFeature feature2 = iterator.next();
-        assertNotNull( feature2 );
-        iterator.close();
+        SimpleFeatureIterator i = features.features();
+        for (SimpleFeature feature = i.next(); i.hasNext(); feature = i.next()) {
+            Property p = feature.getProperty("geom");
+            assertTrue(p.getType() instanceof GeometryType);
+            assertTrue(p.getValue() instanceof Geometry);
+        }
         
-        // Encode |
-        assertEquals("feeling a bit |broken up", feature1.getAttribute("data"));
-        
-        // Encode newline; but still respect trim of white space
-        assertTrue( feature2.getAttribute("geom") instanceof LineString );
-        assertEquals("\nnew line\ndiff friendly", feature2.getAttribute("data"));
     }
-    @Test
-
+    
     public void testSimple() throws Exception {
         SimpleFeatureSource road = store.getFeatureSource( "road" );
         SimpleFeatureCollection features = road.getFeatures();
@@ -230,11 +209,9 @@ public class PropertyDataStore2Test {
         //assertEquals( 1, features.getFeatureType().getAttributeCount() );
         assertEquals( 4, features.size() );
     }
-    @Test
-
     public void testQuery() throws Exception {
         SimpleFeatureSource road = store.getFeatureSource( "road" );
-                
+        
         Query query = new Query( "road", Filter.INCLUDE,
                 new String[]{ "name" } );
         
@@ -242,11 +219,10 @@ public class PropertyDataStore2Test {
         assertEquals( 4, features.size() );
         //assertEquals( 1, features.getFeatureType().getAttributeCount() );
     }
-    @Test
+    
     public void testQueryReproject() throws Exception {
         CoordinateReferenceSystem world = CRS.decode("EPSG:4326"); // world lon/lat
         CoordinateReferenceSystem local = CRS.decode("EPSG:3005"); // british columbia
-        
         
         SimpleFeatureSource road = store.getFeatureSource( "road" );
         SimpleFeatureType origionalType = road.getSchema();
@@ -260,103 +236,78 @@ public class PropertyDataStore2Test {
         SimpleFeatureCollection features = road.getFeatures( query );
         SimpleFeatureType resultType = features.getSchema();
         
-        
         assertNotNull( resultType );
         assertNotSame( resultType, origionalType );
-
+        
         assertEquals( world, resultType.getCoordinateReferenceSystem() );
-
-        GeometryDescriptor geometryDescriptor = resultType.getGeometryDescriptor();        
-        assertTrue( Geometry.class.isAssignableFrom( geometryDescriptor.getType().getBinding() ) );
+        
+        assertNotNull(resultType.getGeometryDescriptor());
     }
+    
+    public void testGetFeaturesFilterSize() throws Exception {
+        Filter f = CQL.toFilter("name = 'brent'");
+        SimpleFeatureSource features = (SimpleFeatureSource) store.getFeatureSource("road");
+        assertEquals(1, features.getFeatures(f).size());
+        
+    }
+    
+    public void testGetFeaturesFilterBounds() throws Exception {
+        Filter f = CQL.toFilter("name = 'brent'");
+        SimpleFeatureSource features = (SimpleFeatureSource) store.getFeatureSource("road");
+        ReferencedEnvelope envelope = new ReferencedEnvelope(20, 30, 20, 30, null);
+        assertEquals(envelope, features.getFeatures(f).getBounds());
+        
+    }
+    
+    /**
+     * Test query with a start index
+     * @throws IOException 
+     * @throws FileNotFoundException 
+     */
     @Test
-    public void testUnusalRoundTrip() throws Throwable {
+    public void testOffset() throws FileNotFoundException, IOException {
+        Query query = new Query(Query.ALL);
+        query.setStartIndex(3);
+        SimpleFeatureSource features = (SimpleFeatureSource) store.getFeatureSource("road");
 
-        File target = new File( "propertyTestData3/trip.properties" );
-        if( target.exists() ){
-            boolean deleted = target.delete();
-            assertTrue( "unable to delete "+target.getAbsolutePath(), deleted );
-        }
-        assertFalse( "trip.properties should not exist yet", target.exists() );
-        
-        
-        SimpleFeatureType schema = DataUtilities.createType("trip", "point:Point::srid=4326,text:String, number:Integer");
-        SimpleFeatureBuilder builder = new SimpleFeatureBuilder( schema );
-        GeometryFactory gf = JTSFactoryFinder.getGeometryFactory();
-        
-        List<SimpleFeature> list = new ArrayList<SimpleFeature>();
-        SimpleFeature feature;
-        
-        feature = builder.buildFeature("trip1", new Object[]{gf.createPoint( new Coordinate(0,0)), "hello world", 1 });
-        feature.getUserData().put( Hints.USE_PROVIDED_FID, true );
-        list.add( feature);
-        
-        feature = builder.buildFeature("trip2", new Object[]{gf.createPoint( new Coordinate(0,0)), "test if | chracter handling", 2 });
-        feature.getUserData().put( Hints.USE_PROVIDED_FID, true );
-        list.add( feature );
-        
-        feature= builder.buildFeature("trip3", new Object[]{gf.createPoint( new Coordinate(0,0)), "test of\n multi-line handling", 3 });
-        feature.getUserData().put( Hints.USE_PROVIDED_FID, true );
-        list.add( feature );
+        SimpleFeatureCollection matches = features.getFeatures(query);
 
-        feature = builder.buildFeature("trip4", new Object[]{gf.createPoint( new Coordinate(0,0)), "    test of\n whitespace handling", 4 });
-        feature.getUserData().put( Hints.USE_PROVIDED_FID, true );
-        list.add( feature );
-        
-        feature = builder.buildFeature("trip5", new Object[]{gf.createPoint( new Coordinate(0,0)), "test encoding does not get confused over \\n newline and \\twhite space and \\| other markers", 5 });
-        feature.getUserData().put( Hints.USE_PROVIDED_FID, true );
-        list.add( feature );
-        
-        feature = builder.buildFeature("trip6", new Object[]{gf.createPoint( new Coordinate(0,0)), "How well can we encode 1\\2?", 5 });
-        feature.getUserData().put( Hints.USE_PROVIDED_FID, true );
-        list.add( feature );
-        
-        ListFeatureCollection features = new ListFeatureCollection(schema, list );
-        
-        unusalStore.createSchema( schema );
-        SimpleFeatureStore trip = (SimpleFeatureStore) unusalStore.getFeatureSource("trip");
-        
-        trip.addFeatures( features );
-        assertTrue( "trip.properties created", target.exists() );
-        
-        // listFileContents(target);
-        
-        assertEquals("stored", list.size(), trip.getCount(Query.ALL));
-        
-        final Map<String,SimpleFeature> cache = new HashMap<String,SimpleFeature>();
-        SimpleFeatureCollection readFeatures = trip.getFeatures();
-        
-        FeatureVisitor cacheResults = new FeatureVisitor() {
-            @Override
-            public void visit(Feature f) {
-                SimpleFeature feature = (SimpleFeature) f;
-                cache.put( feature.getID(), feature );
-            }
-        };
-        readFeatures.accepts( cacheResults, null );
-        assertEquals( "restored", list.size(), cache.size() );
-        assertEquals( "hello world", cache.get("trip1").getAttribute("text"));
-        assertEquals( "test if | chracter handling", cache.get("trip2").getAttribute("text"));
-        assertEquals( "test of\n multi-line handling", cache.get("trip3").getAttribute("text"));
-        assertEquals( "    test of\n whitespace handling", cache.get("trip4").getAttribute("text"));
-        
-        // need some regex magic to make this work
-        //assertEquals( "test encoding does not get confused over \\n newline and \\twhite space and \\| other markers", cache.get("trip5").getAttribute("text"));
-        assertEquals( "How well can we encode 1\\2?", cache.get("trip6").getAttribute("text"));
+        assertEquals(1, matches.size());
+        assertEquals(1, features.getCount(query));
     }
+    
+    /**
+     * Test query with maxFeatures
+     * @throws IOException 
+     * @throws FileNotFoundException 
+     */
+    @Test
+    public void testLimit() throws FileNotFoundException, IOException {
+        Query query = new Query(Query.ALL);
+        query.setMaxFeatures(3);
+        SimpleFeatureSource features = (SimpleFeatureSource) store.getFeatureSource("road");
 
-    private void listFileContents(File target) throws FileNotFoundException, IOException {
-        System.out.println("Contents of "+target );
-        if( target.exists() ){
-            BufferedReader reader = new BufferedReader( new FileReader( target ) );
-            String line;
-            while( (line = reader.readLine()) != null ){
-                System.out.println( line );
-            }
-            reader.close();
-        }
-        else {
-            System.out.println(" ... does not exist");
-        }
+        SimpleFeatureCollection matches = features.getFeatures(query);
+
+        assertEquals(3, matches.size());
+        assertEquals(3, features.getCount(query));
+    }
+    
+    /**
+     * Test query with maxFeatures and startIndex
+     * @throws IOException 
+     * @throws FileNotFoundException 
+     */
+    @Test
+    public void testLimitOffset() throws FileNotFoundException, IOException {
+        Query query = new Query(Query.ALL);
+        query.setMaxFeatures(3);
+        query.setStartIndex(3);
+        SimpleFeatureSource features = (SimpleFeatureSource) store.getFeatureSource("road");
+
+        SimpleFeatureCollection matches = features.getFeatures(query);
+
+        assertEquals(1, matches.size());
+        assertEquals(1, features.getCount(query));
     }
 }
