+++ modules/unsupported/pom.xml
@@ -161,7 +161,7 @@
          <module>sfs</module>
          <module>feature-aggregate</module>
          <module>mongodb</module>
-         <module>property-ng</module>
+         <module>property-old</module>
          <module>solr</module>
          <module>css</module>
        </modules>
