--- modules/unsupported/property-ng/src/main/resources/META-INF/services/org.geotools.data.DataStoreFactorySpi
@@ -1,1 +1,0 @@
-org.geotools.data.property.ng.PropertyDataStoreFactory
