+++ modules/unsupported/property-old/src/main/resources/META-INF/services/org.geotools.data.DataStoreFactorySpi
@@ -1,0 +1,1 @@
+org.geotools.data.property.old.PropertyDataStoreFactory
